add: restrictedAccess data

+++ server/src/controllers/contacts/removeById.js
@@ -2,9 +2,12 @@ const { HttpError } = require('../../utils');
 
 const { Contact } = require('../../models');
 const { ctrlWrapper } = require('../../decorators');
+const { restrictedAccess } = require('../../utils');
 
 const removeById = ctrlWrapper(async (req, res) => {
   const { id } = req.params;
+  if (restrictedAccess.contactId.includes(id)) throw HttpError(403);
+
   const deletedContact = await Contact.findByIdAndDelete(id);
   if (!deletedContact) throw HttpError(403);
